kahoot_auto_answer/requirements.txt
mss
Pillow
pytesseract
pynput
google-generativeai
pyautogui
pandas